# shared/openapi.yaml
openapi: 3.0.3
info:
  title: ExplorArte API
  version: 1.0.0
  description: >
    REST contract for the ExplorArte platform (Sueños y Letras). Both the React
    web app and the Expo mobile app talk to this API through the shared ApiClient.
    Until a server exists they run against an in-memory mock that implements the
    same shapes defined here. Implement this contract in Python (FastAPI) or Java
    (Spring Boot); the @explorarte/shared seed module provides initial data.
servers:
  - url: http://localhost:8000
    description: Local backend
  - url: https://api.explorarte.example
    description: Production
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Email + password login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401': { description: Invalid credentials }
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
  /auth/invitations/{token}:
    get:
      tags: [Auth]
      summary: Check an admin invitation token (public — opened before signing up)
      description: >
        A token that is unknown, expired, already used or revoked all answer the
        same `{ valid: false }`, so this cannot be used to probe for valid tokens.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid: { type: boolean }
                  email: { type: string, nullable: true }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Send a password reset link / code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emailOrPhone]
              properties:
                emailOrPhone: { type: string }
      responses:
        '200':
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean, enum: [true] }
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, lastname, email, phone, institucion, ubicacion, role, status]
      properties:
        id: { type: string }
        name: { type: string }
        lastname: { type: string }
        email: { type: string }
        phone: { type: string }
        institucion: { type: string }
        ubicacion: { type: string }
        role: { type: string, enum: [teacher, admin] }
        status: { type: string, enum: [pending, approved, rejected] }
        photo: { type: string, nullable: true }
        profileCompleted:
          type: boolean
          description: >
            False while an invited account has not completed its profile. Drives
            the welcome notice; flips to true when the profile is saved with a
            name, a lastname and a location.
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    RegisterInput:
      type: object
      required: [name, lastname, email, password]
      properties:
        name: { type: string }
        lastname: { type: string }
        ubicacion: { type: string }
        email: { type: string }
        password: { type: string, minLength: 8 }
        invitationToken:
          type: string
          maxLength: 200
          description: Present when the signup came from an admin invitation email.
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/UserProfile' }